requirements.txt
anyio==3.7.1
bardapi==0.1.30
beautifulsoup4==4.12.2
beepy==1.0.7
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comtypes==1.2.0
deep-translator==1.11.4
docopt==0.6.2
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.11.3
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
Js2Py==0.74
lz4==4.3.2
packaging==23.1
pipwin==0.5.2
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycryptodomex==3.18.0
pyjsparser==2.7.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
requests==2.31.0
rsa==4.9
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
